Extract settings methods into `Meilisearch::Index::Settings` module
